fix: ROUND(double precision, integer) 需显式转 NUMERIC

+++ xhs-cloud/cloud_deploy/scripts/compute_feature_metrics.py
@@ -137,7 +137,7 @@ def run_compute(*, dry_run: bool = False, snap_date: str | None = None) -> dict:
         UPDATE goods_feature_metrics cur
         SET acceleration = CASE
                 WHEN COALESCE(prev.growth_rate, 0) != 0 AND cur.growth_rate != 0
-                THEN ROUND(cur.growth_rate - prev.growth_rate, 6)
+                THEN ROUND((cur.growth_rate - prev.growth_rate)::NUMERIC, 6)
                 ELSE 0
             END,
             updated_at = to_char(NOW(), 'YYYY-MM-DD HH24:MI:SS')
